print sizes

+++ hello_world/6-size.c
@@ -15,7 +15,7 @@ int main(void)
 	printf("The size of an int is: %lu byte(s)\n", (unsigned long)sizeof(i));
 	printf("The size of a long int is: %lu byte(s)\n", (unsigned long)sizeof(li));
 	printf("The size of a long long int is: %lu byte(s)\n", (unsigned long)sizeof(lli));
-	printf("The size of a float is: %lu byte(s)\n", (unsigned long)sizeof(f));
+	printf("The size of a float is: %lu byte(s)\n", sizeof(f));
 	
-	return(0);
+	return (0);
 }
